order更新

+++ gm_goods/gm_goods_service/src/main/java/com/gomai/goods/service/Category1Service.java
@@ -5,5 +5,5 @@ import com.gomai.goods.pojo.Category1;
 import java.util.List;
 
 public interface Category1Service {
-    public List<Category1> queryListByParent(Long ca1_id);
+    public List<Category1> queryListByParent(Integer ca1_id);
 }
